refactor: WAS와 Web application의 패키지구조 변경
