+++ src/main/java/web/application/controller/Controller.java
@@ -1,0 +1,9 @@
+package web.application.controller;
+
+import web.server.domain.request.HttpRequest;
+import web.server.domain.response.HttpResponse;
+
+public interface Controller {
+
+    void service(HttpRequest httpRequest, HttpResponse httpResponse);
+}
