+++ src/main/java/web/application/controller/StaticController.java
@@ -1,0 +1,27 @@
+package web.application.controller;
+
+import web.server.domain.request.HttpRequest;
+import web.server.domain.response.HttpResponse;
+
+public class StaticController extends AbstractController {
+
+    private StaticController() {
+    }
+
+    public static StaticController getInstance() {
+        return ControllerCache.staticController;
+    }
+
+    @Override
+    public void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
+        httpResponse.forward(httpRequest.getPath());
+    }
+
+    @Override
+    public void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
+    }
+
+    private static class ControllerCache {
+        private static final StaticController staticController = new StaticController();
+    }
+}
