+++ src/main/java/web/server/exception/UrlNotFoundException.java
@@ -1,0 +1,8 @@
+package web.server.exception;
+
+public class UrlNotFoundException extends IllegalArgumentException {
+
+    public UrlNotFoundException(String path) {
+        super("등록되지 않은 경로입니다. Path = " + path);
+    }
+}
