--- src/main/java/webserver/controller/Controller.java
@@ -1,9 +1,0 @@
-package webserver.controller;
-
-import webserver.domain.request.HttpRequest;
-import webserver.domain.response.HttpResponse;
-
-public interface Controller {
-
-    void service(HttpRequest httpRequest, HttpResponse httpResponse);
-}
