+++ src/test/java/resources/http_get2.txt
@@ -1,0 +1,9 @@
+GET /user/create?userId=javajigi&password=password&name=JaeSung HTTP/1.1
+Host: localhost:8080
+Connection: keep-alive
+Content-Length: 4
+Content-Type: application/x-www-form-urlencoded
+Accept: */*
+
+id=1
+
