+++ src/test/java/web/server/utils/HandlebarsTest.java
@@ -1,0 +1,29 @@
+package web.server.utils;
+
+import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.github.jknack.handlebars.Handlebars;
+import com.github.jknack.handlebars.Template;
+import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
+import com.github.jknack.handlebars.io.TemplateLoader;
+import web.application.domain.model.User;
+
+public class HandlebarsTest {
+    private static final Logger log = LoggerFactory.getLogger(HandlebarsTest.class);
+
+    @Test
+    void name() throws Exception {
+        TemplateLoader loader = new ClassPathTemplateLoader();
+        loader.setPrefix("/templates");
+        loader.setSuffix(".html");
+        Handlebars handlebars = new Handlebars(loader);
+
+        Template template = handlebars.compile("user/profile");
+
+        User user = new User("javajigi", "password", "자바지기", "javajigi@gmail.com");
+        String profilePage = template.apply(user);
+        log.debug("ProfilePage : {}", profilePage);
+    }
+}
