+++ src/test/java/web/server/utils/IOUtilsTest.java
@@ -1,0 +1,21 @@
+package web.server.utils;
+
+import java.io.BufferedReader;
+import java.io.StringReader;
+
+import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class IOUtilsTest {
+    private static final Logger logger = LoggerFactory.getLogger(IOUtilsTest.class);
+
+    @Test
+    public void readData() throws Exception {
+        String data = "abcd123";
+        StringReader sr = new StringReader(data);
+        BufferedReader br = new BufferedReader(sr);
+
+        logger.debug("parse body : {}", IOUtils.readData(br, data.length()));
+    }
+}
